+++ changelog.rst
@@ -1,6 +1,11 @@
 ChangeLog
 =========
 
+1.4.2 (2015/02/22)
+------------------
+
+- add requirements.txt and changelog.rst in building packages.
+
 1.4.1 (2015/02/21)
 ------------------